requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
dj-static==0.0.6
Django==4.2.7
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
pygraphviz==1.11
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
Pillow==10.1.0
#  wsgiref